requirements.txt
openai
librosa
pydub
silero_vad
soundfile
tqdm
setuptools
wheel
requests
srt
fastapi
uvicorn
python-multipart
websockets
python-dotenv

# Optional (for embedded qwen_asr upstream demos):
# transformers==4.57.6
# nagisa==0.2.11
# soynlp==0.0.493
# accelerate==1.12.0
# qwen-omni-utils
# sox
# pytz
# gradio
# flask
# scipy
